refactor: organizando os diretorios do back-end
