+++ RPP/src/br/rpp/ficha/Caracteristica.java
@@ -1,6 +1,5 @@
-package br.rpp;
+package br.rpp.ficha;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 public class Caracteristica {
